Disable jekyll for GitHub Pages
We don't need the extra overhead for Jekyll, and Jekyll is doing some
things (like ignoring hidden files/folders) that we don't actually want
it to do. Turning off Jekyll seems like the best thing to do.

To do this, we need a .nojekyll file at the root of the repo. Hakyll
hates hidden files (see .well-known), so we'll have to create a specific
matcher for it and turn a visible file into a hidden one manually.
Unfortunate.