Add function to replace relative urls
When we're in a feed, we need to be able to link to absolute urls.
However, it's nice to be able to refer to relative urls when writing or
when on our own domain.

To fix this, we can build an optional compiler step that replaces
relative urls with absolute ones.

The way we do this is to look for any quoted strings that appear to be
assignments that start with `/` (literally `="/`). This is probably
pretty naive, but we'll deal with that later. If we find that, we drop
those characters, then prepend the result with the provided host name
(using `</>` to preserve the slash), then replace the beginning `="`.