Add function to repair external urls
We're kind of abusing Hakyll's support for urls. In our stubs, we're
redefining the url to be external. Hakyll doesn't expect this, which
isn't a huge deal if we're doing things locally, but it's also trying to
help us when it constructs links for the RSS feed. In that case, it
prepends the url with the site host. Which, obviously, isn't good for
us, since those links don't exist on our domain.

To fix this, we can look for links that start with the provided host and
then immediately have `http` after them ("http://example.comhttp").
These are obviously malformed urls. We can then just drop the characters
in the provided host, and voila, we're left with the original url.