Customize build process
The `site` binary is actually _super_ useful, and has some interesting
shit in it. Plus, this is a really good opportunity to learn more about
cabal.

This process is being cribbed from
https://github.com/blaenk/blaenk.github.io/blob/source/Setup.hs

I've renamed some things and broken some things up in an effort to grok
what was going on. Essentially:

 - We look at the build package description and make sure we have an
   executable named 'site'
 - if so, we get the build directory from the build info, and copy the
   'site' binary out of there, and into the pwd.

This will allow us to run `site` directly without having to go through
cabal or dig through the `dist` folder.