Clean up indexed route for blog posts
We were previously doing wacky shit like removing the extension and
composing that with the indexedRoute, but that didn't make sense since
that route was super specialized for blog posts anyway. Instead, we
should just move the extension stuff into the route itself by only
taking the filename after dropping the date prefix.