Add Years module
This module is based _heavily_ off the Tags module provided by Hakyll.
It allows you to interact with years the same way you'd interact with
tags.

It expects to have posts with the standard format of
`YYYY-MM-DD-Name.markdown`. It pulls the year from the filename and
constructs an index page at the specified path with the posts matching
that year.

I _barely_ understand what's going on here.