Compile sass files
This adds a new compiler that is able to handle compiling sass. We're
only using it to compile a single file (screen.scss), which will end up
being our main scss file. This file will import the other files that we
need, allowing us to break up the file semantically.

This is Really Cool Stuff IMO.