Rebuild screen.scss if any other sass file changes
The problem with the previous approach is that once we break up the scss
into multiple files, Hakyll doesn't know about the sub-files, so it
doesn't rebuild the larger scss file when one of them changes.

To fix this, we can use the moderately confusing
`rulesExtraDependencies` function. First, we create a dependency with a
pattern (in this case, any scss files in the scss directory). This will
tell Hakyll to pay attention to those files. Then, we can pass those
into `rulesExtraDependencies` and then pass our normal build rule into
that.

The end result is that if any of those files change, the cache is
invalidated, and it will re-run the sass compiler for `screen.scss`.