Simplify website structure
I don't like the projects page, lets kill it.

Now that we're down to one page, lets go ahead and just link my name to /about
and back to / if we're already on /about. This means flipping out the
template, which seems fine tbh. A little duplication, but not too bad.