Create custom field for getting today's date
This field is similar to the `dateField` except that it gets the date from the
current date/time on the system instead of looking at the file.

This is going to be useful for when we need to give a date to our draft blog
posts, which don't have one automatically.