Only perform draft functions when watching
Instead of always including drafts, we should only use the draft functionality
if we're running `site watch`, which is analogous to `yesod devel`. When
watching, we want to provide the default date for all posts, and compile our
drafts along with our posts.

This diff is much smaller than it would appear. Most of this is re-indenting
the `main` function to accommodate the new `do` block.